Move diagnosis code into IntegrationTest
